requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.0.0
pyTelegramBotAPI
requests
